requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
babel==2.15.0
    # via docxcompose
blinker==1.8.2
    # via flask
boto3==1.34.144
botocore==1.34.144
    # via
    #   boto3
    #   s3transfer
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via cryptography
chardet==5.2.0
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
cryptography==43.0.0
    # via
    #   msal
    #   pyjwt
docx-mailmerge==0.5.0
docxcompose==1.4.0
    # via docxtpl
docxtpl==0.17.0
flask==3.0.3
    # via
    #   flask-login
    #   flask-wtf
flask-login==0.6.3
flask-wtf==1.2.1
greenlet==3.0.3
    # via sqlalchemy
gunicorn==22.0.0
idna==3.7
    # via requests
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.4
    # via
    #   docxtpl
    #   flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
loguru==0.7.2
lxml==5.2.2
    # via
    #   docx-mailmerge
    #   docxcompose
    #   docxtpl
    #   python-docx
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
msal==1.30.0
    # via office365-rest-python-client
mysql-connector-python==9.0.0
numpy==2.0.0
    # via pandas
office365-rest-python-client==2.5.11
packaging==24.1
    # via gunicorn
pandas==2.2.2
pendulum==3.0.0
psycopg2-binary==2.9.9
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via msal
pymysql==1.1.1
pyodbc==5.1.0
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
    #   pendulum
    #   time-machine
python-docx==1.1.2
    # via
    #   docxcompose
    #   docxtpl
python-dotenv==1.0.1
pytz==2024.1
    # via
    #   office365-rest-python-client
    #   pandas
pyyaml==6.0.1
regex==2024.5.15
requests==2.32.3
    # via
    #   msal
    #   office365-rest-python-client
s3transfer==0.10.2
    # via boto3
setuptools==70.1.0
    # via docxcompose
six==1.16.0
    # via
    #   docxcompose
    #   docxtpl
    #   python-dateutil
sqlalchemy==2.0.31
time-machine==2.14.1
    # via pendulum
typing-extensions==4.12.2
    # via
    #   office365-rest-python-client
    #   python-docx
    #   sqlalchemy
tzdata==2024.1
    # via
    #   pandas
    #   pendulum
urllib3==2.2.2
    # via
    #   botocore
    #   requests
werkzeug==3.0.3
    # via
    #   flask
    #   flask-login
wtforms==3.1.2
    # via flask-wtf
